refactor: Rename Displays to TempDisplay
